% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradDescent.R
\name{grad_descent}
\alias{grad_descent}
\title{Finds the location of a local minimum of a function using gradient descent.}
\usage{
grad_descent(f, x, grad_f)
}
\arguments{
\item{f}{A real valued function.}

\item{x}{A initial estimate for the local minimum.}

\item{grad_f}{The gradient of f (Optional).}
}
\value{
A local minimum of the function f.
}
\description{
Finds the location of a local minimum of a function using gradient descent.
}
