% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradDescent.R
\name{gradDescent}
\alias{gradDescent}
\title{Finds the location of a local minimum of a function using gradient descent.}
\usage{
gradDescent(f)
}
\arguments{
\item{f}{A single valued function.}
}
\value{
A local minimum of the function f.
}
\description{
Finds the location of a local minimum of a function using gradient descent.
}
