% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochasticGradientDescent.R
\name{stoc_grad_descent}
\alias{stoc_grad_descent}
\title{Finds the location of a local minimum of a function using stochastic gradient descent.}
\usage{
stoc_grad_descent(f, data, w, step_size = 1e-04)
}
\arguments{
\item{f}{A real valued differentiable function}

\item{data}{The data set to b}

\item{w}{Initial guess for a local minimum of f}
}
\value{
An estimate for a local minimum of \sum f.
}
\description{
Finds the location of a local minimum of a function using stochastic gradient descent.
}
